+++ config/onboarding.php
@@ -11,24 +11,94 @@
  */
 
 return array(
-	'dependencies' => array(
+	'dependencies'     => array(
 		'plugins' => array(
 			array(
-				'name'        => __( 'Atomic Blocks', 'genesis-sample' ),
-				'slug'        => 'atomic-blocks/atomicblocks.php',
+				'name'       => __( 'Atomic Blocks', 'genesis-sample' ),
+				'slug'       => 'atomic-blocks/atomicblocks.php',
+				'public_url' => 'https://atomicblocks.com/',
+			),
+			array(
+				'name'       => __( 'WP Forms Lite', 'genesis-sample' ),
+				'slug'       => 'wpforms-lite/wpforms.php',
+				'public_url' => 'https://wordpress.org/plugins/wpforms-lite/',
+			),
+			array(
+				'name'       => __( 'Genesis eNews Extended', 'genesis-sample' ),
+				'slug'       => 'genesis-enews-extended/plugin.php',
+				'public_url' => 'https://wordpress.org/plugins/genesis-enews-extended/',
+			),
+			array(
+				'name'       => __( 'Simple Social Icons', 'genesis-sample' ),
+				'slug'       => 'simple-social-icons/simple-social-icons.php',
+				'public_url' => 'https://wordpress.org/plugins/simple-social-icons/',
 			),
 		),
 	),
-	'content' => array(
+	'content'          => array(
 		'homepage' => array(
 			'post_title'     => 'Homepage',
-			'post_name'      => 'homepage-gutenberg',
 			'post_content'   => require dirname( __FILE__ ) . '/import/content/homepage.php',
 			'post_type'      => 'page',
 			'post_status'    => 'publish',
-			'page_template'  => 'template-blocks.php',
+			'page_template'  => 'page-templates/blocks.php',
 			'comment_status' => 'closed',
 			'ping_status'    => 'closed',
+		),
+		'blocks'   => array(
+			'post_title'     => 'Block Content Examples',
+			'post_content'   => require dirname( __FILE__ ) . '/import/content/block-examples.php',
+			'post_type'      => 'page',
+			'post_status'    => 'publish',
+			'page_template'  => 'page-templates/blocks.php',
+			'comment_status' => 'closed',
+			'ping_status'    => 'closed',
+		),
+		'about'    => array(
+			'post_title'     => 'About Us',
+			'post_content'   => require dirname( __FILE__ ) . '/import/content/about.php',
+			'post_type'      => 'page',
+			'post_status'    => 'publish',
+			'page_template'  => 'page-templates/blocks.php',
+			'featured_image' => CHILD_URL . '/config/import/images/about.jpg',
+			'comment_status' => 'closed',
+			'ping_status'    => 'closed',
+		),
+		'contact'  => array(
+			'post_title'     => 'Contact Us',
+			'post_content'   => require dirname( __FILE__ ) . '/import/content/contact.php',
+			'post_type'      => 'page',
+			'post_status'    => 'publish',
+			'comment_status' => 'closed',
+			'ping_status'    => 'closed',
+		),
+		'landing'  => array(
+			'post_title'     => 'Landing Page',
+			'post_content'   => require dirname( __FILE__ ) . '/import/content/landing-page.php',
+			'post_type'      => 'page',
+			'post_status'    => 'publish',
+			'page_template'  => 'page-templates/landing.php',
+			'comment_status' => 'closed',
+			'ping_status'    => 'closed',
+		),
+	),
+	'navigation_menus' => array(
+		'primary' => array(
+			'homepage' => array(
+				'title' => 'Home',
+			),
+			'about'    => array(
+				'title' => 'About Us',
+			),
+			'contact'  => array(
+				'title' => 'Contact Us',
+			),
+			'blocks'   => array(
+				'title' => 'Block Examples',
+			),
+			'landing'  => array(
+				'title' => 'Landing Page',
+			),
 		),
 	),
 );