+++ lib/gutenberg/init.php
@@ -16,12 +16,10 @@ add_action( 'wp_enqueue_scripts', 'genesis_sample_enqueue_gutenberg_frontend_sty
  */
 function genesis_sample_enqueue_gutenberg_frontend_styles() {
 
-	$child_theme_slug = defined( 'CHILD_THEME_NAME' ) && CHILD_THEME_NAME ? sanitize_title_with_dashes( CHILD_THEME_NAME ) : 'genesis-sample';
-
 	wp_enqueue_style(
 		'genesis-sample-gutenberg',
 		get_stylesheet_directory_uri() . '/lib/gutenberg/front-end.css',
-		array( $child_theme_slug ),
+		array( CHILD_THEME_HANDLE ),
 		CHILD_THEME_VERSION
 	);
 
@@ -35,9 +33,11 @@ add_action( 'enqueue_block_editor_assets', 'genesis_sample_block_editor_styles' 
  */
 function genesis_sample_block_editor_styles() {
 
+	$block_editor_settings = genesis_get_config( 'block-editor-settings' );
+
 	wp_enqueue_style(
 		'genesis-sample-gutenberg-fonts',
-		'https://fonts.googleapis.com/css?family=Source+Sans+Pro:400,400i,600,700',
+		$block_editor_settings['admin-fonts-url'],
 		array(),
 		CHILD_THEME_VERSION
 	);
@@ -59,7 +59,7 @@ add_filter( 'body_class', 'genesis_sample_blocks_body_classes' );
  */
 function genesis_sample_blocks_body_classes( $classes ) {
 
-	if ( ! is_singular() || ! function_exists( 'has_blocks' ) || ! function_exists( 'parse_blocks') ) {
+	if ( ! is_singular() || ! function_exists( 'has_blocks' ) || ! function_exists( 'parse_blocks' ) ) {
 		return $classes;
 	}
 
@@ -95,17 +95,21 @@ add_theme_support( 'align-wide' );
 // Make media embeds responsive.
 add_theme_support( 'responsive-embeds' );
 
+$genesis_sample_block_editor_settings = genesis_get_config( 'block-editor-settings' );
+
 // Adds support for editor font sizes.
 add_theme_support(
 	'editor-font-sizes',
-	genesis_get_config( 'editor-font-sizes' )
+	$genesis_sample_block_editor_settings['editor-font-sizes']
 );
 
 // Adds support for editor color palette.
 add_theme_support(
 	'editor-color-palette',
-	genesis_get_config( 'editor-color-palette' )
+	$genesis_sample_block_editor_settings['editor-color-palette']
 );
+
+require_once get_stylesheet_directory() . '/lib/gutenberg/inline-styles.php';
 
 add_action( 'after_setup_theme', 'genesis_sample_content_width', 0 );
 /**
@@ -113,7 +117,9 @@ add_action( 'after_setup_theme', 'genesis_sample_content_width', 0 );
  */
 function genesis_sample_content_width() {
 
+	$block_editor_settings = genesis_get_config( 'block-editor-settings' );
+
 	// phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedVariableFound -- See https://github.com/WordPress-Coding-Standards/WordPress-Coding-Standards/issues/924
-	$GLOBALS['content_width'] = apply_filters( 'genesis_sample_content_width', 1062 );
+	$GLOBALS['content_width'] = apply_filters( 'genesis_sample_content_width', $block_editor_settings['content-width'] );
 
 }