+++ lib/output.php
@@ -19,8 +19,6 @@ add_action( 'wp_enqueue_scripts', 'genesis_sample_css' );
  */
 function genesis_sample_css() {
 
-	$handle = defined( 'CHILD_THEME_NAME' ) && CHILD_THEME_NAME ? sanitize_title_with_dashes( CHILD_THEME_NAME ) : 'child-theme';
-
 	$color_link   = get_theme_mod( 'genesis_sample_link_color', genesis_sample_customizer_get_default_link_color() );
 	$color_accent = get_theme_mod( 'genesis_sample_accent_color', genesis_sample_customizer_get_default_accent_color() );
 	$logo         = wp_get_attachment_image_src( get_theme_mod( 'custom_logo' ), 'full' );
@@ -73,6 +71,10 @@ function genesis_sample_css() {
 		input[type="reset"]:hover,
 		input[type="submit"]:focus,
 		input[type="submit"]:hover,
+		.site-container div.wpforms-container-full .wpforms-form input[type="submit"]:focus,
+		.site-container div.wpforms-container-full .wpforms-form input[type="submit"]:hover,
+		.site-container div.wpforms-container-full .wpforms-form button[type="submit"]:focus,
+		.site-container div.wpforms-container-full .wpforms-form button[type="submit"]:hover,
 		.button:focus,
 		.button:hover {
 			background-color: %1$s;
@@ -145,7 +147,7 @@ function genesis_sample_css() {
 	) : '';
 
 	if ( $css ) {
-		wp_add_inline_style( $handle, $css );
+		wp_add_inline_style( CHILD_THEME_HANDLE, $css );
 	}
 
 }